Remove GoogleMaps js API code replaced by OpenLayers
