+++ .github/workflows/release.yaml
@@ -32,8 +32,6 @@ jobs:
         run: |
           cat << EOF > .env
           VITE_APP_VERSION=${{ github.event.release.tag_name }}
-          VITE_APP_GOOGLE_MAPS_API_KEY=${{ secrets.GOOGLE_MAPS_API_KEY }}
-          VITE_APP_GOOGLE_MAPS_MAP_ID=${{ secrets.GOOGLE_MAPS_MAP_ID }}
           VITE_APP_PROXY_BASE_URL=${{ vars.PROXY_BASE_URL }}
           EOF
 
@@ -68,6 +66,6 @@ jobs:
           GITHUB_TOKEN: ${{ secrets.GITHUB_TOKEN }}
         with:
           upload_url: ${{ github.event.release.upload_url }}
-          asset_path: "data-management-app-${{ github.event.release.tag_name }}.zip"
-          asset_name: "data-management-app-${{ github.event.release.tag_name }}.zip"
+          asset_path: 'data-management-app-${{ github.event.release.tag_name }}.zip'
+          asset_name: 'data-management-app-${{ github.event.release.tag_name }}.zip'
           asset_content_type: application/zip
