--- src/config/googleMapsConfig.ts
@@ -1,5 +1,0 @@
-// https://developers.google.com/maps/documentation/javascript/reference/map#MapOptions
-export const defaultMapOptions = {
-  center: { lat: 39, lng: -100 },
-  zoom: 4,
-}
