+++ src/pages/Sites.vue
@@ -137,7 +137,7 @@ import SiteFilterToolbar from '@/components/Site/SiteFilterToolbar.vue'
 import WorkspaceToolbar from '@/components/Workspace/WorkspaceToolbar.vue'
 import { api } from '@/services/api'
 import { PermissionResource, PermissionAction, Thing } from '@/types'
-import { addColorToMarkers } from '@/utils/googleMaps/markers'
+import { addColorToMarkers } from '@/utils/maps/markers'
 import { ThingWithColor } from '@/types'
 import { Snackbar } from '@/utils/notifications'
 import { storeToRefs } from 'pinia'
